+++ src/engine/settingsengine.cpp
@@ -74,7 +74,7 @@ void SettingsEngine::initialize() {
     std::vector<std::string> scenes = ghoul::filesystem::Directory(sceneDir).readFiles();
     for (std::size_t i = 0; i < scenes.size(); ++i) {
         std::size_t found = scenes[i].find_last_of("/\\");
-        _scenes.addOption(i, scenes[i].substr(found + 1));
+        _scenes.addOption(static_cast<int>(i), scenes[i].substr(found + 1));
     }
 
     // Set interaction to change ConfigurationManager and schedule the load
